basic types and operators
Basic types and operators of the Java programming language.
Creating a class that consists of an executive method that calculates the value of a function with the "float" index type, "-" and "+" operations, and the constant C = 1. The result of the action is a single value of a real type. All exceptions that may occur during program code execution must be handled. All variables must be described and their values set in the executive method.