Arrays in the Java programming language
Creating a class that consists of an executive method that performs an action on a matrix(s) with an element type of char and an action on the resulting matrix "Calculate the sum of the largest elements in the rows of a matrix with even numbers and the smallest elements in the rows of a matrix with odd numbers". Display the results of the first and second actions. All exceptions that may occur during the execution of the program code must be handled. All variables must be described and their values set in the executive method.