Strings in the Java programming language
Create a class that consists of an executive method that performs an action (In each word of the given text, delete all previous occurrences of the last letter of this word.) with a text string of type StringBuffer. All exceptions that may occur during the execution of the program code must be handled. All variables must be described and their values set in the executable method.