Classes in the Java programming language
Create a class with an executable method in which to create an array of class objects (a boat consisting of at least 5 fields). Sort the array, by one of the fields, in ascending order, and by the other, in descending order, using standard sorting tools. All variables must be described and their values set in the executive method. The code must comp ly withJCCstandards  and be documented in detail.