the relationship between classes
Modify lab3 as follows: create separate classes for letters, words, sentences, punctuation marks, and text. A word should consist of an array of letters, a sentence of an array of words and punctuation marks, and a text of an array of sentences. Replace the sequence of tabs and spaces with a single space.
Create a class that consists of an executive method that performs the described action from lab3, but uses the created classes as types. You must handle all exceptions that may occur during the execution of the program code. All variables must be described and their values set in the executive method. The code must comp ly withJCCstandards  and be documented in detail.