Inheritance and polymorphism
Create a generic class and at least 3 descendant classes that describe a given area of knowledge (Define a hierarchy of airplanes. Create an airline company. Calculate the total capacity and carrying capacity of the airline. Sort the company's aircraft by flight range. Find an airplane in the company that meets the specified range of fuel consumption). Create a class consisting of an array of objects with which you can perform the actions specified by the option. It is necessary to handle all exceptional situations that may arise during the execution of the program code. All variables must be described and their values set in the executive method. The code must comp ly withJCCstandards  and be documented in detail.